requirements.txt
tornado>=6.3.2
pyarrow>=10.0.1
snowflake_snowpark_python>=1.5
pandas>=1.5.3
streamlit>=1.10.0
keplergl==0.2.2
streamlit_keplergl>=0.3.0
folium
geojson
streamlit_folium
streamlit-option-menu
charset-normalizer==3.1.0